refactor(derive): split monolithic lib.rs into per-macro modules
Split 640-line lib.rs into 5 focused modules:
- lib.rs: Crate root with macro entry points and documentation (170 lines)
- attributes.rs: FieldConfig/StructConfig parsing from #[mssql(...)]
- naming.rs: Case conversion utilities (snake_case, camelCase, etc.)
- from_row.rs: FromRow derive implementation
- to_params.rs: ToParams derive implementation
- tvp.rs: Tvp derive implementation with SQL type inference

Each derive macro is now isolated in its own file, making it easier
to understand, modify, and extend. Shared attribute parsing and naming
utilities are cleanly separated.

Resolves STRUCT-003.
